Updated file name and add home page section2 cards
