+++ frontend/src/components/Home.js
@@ -1,10 +1,13 @@
 import React from "react";
 import Section1 from "./Sections/HomeSection1";
+import Section2 from "./Sections/HomeSection2";
 
 const Home = () => {
 	return (
-		<Section1>
-		</Section1>
+		<div>
+			<Section1 />
+			<Section2 />
+		</div>
 	);
 };
