--- frontend/src/components/Sections/HomeSection.module.css
@@ -1,11 +1,0 @@
-.grid_display {
-	display: grid;
-	grid-template-columns: repeat(10, 10%);
-    grid-template-rows: repeat(10, 10%);
-}
-
-.section1 {
-	width: 100%;
-	height: 70vh;
-	background-color: #14274e;;
-}
