+++ frontend/src/components/Sections/HomeSection1.module.css
@@ -1,0 +1,20 @@
+.grid_display {
+	display: grid;
+	grid-template-columns: repeat(10, 10%);
+    grid-template-rows: repeat(10, 10%);
+}
+
+.section1 {
+	width: 100%;
+	height: 70vh;
+	background-color: #14274e;;
+}
+
+.section1_img {
+	background-image: url("../img/hut_bg_2.png");
+	grid-row-start: 1;
+	grid-row-end: -1;
+	grid-column-start: 4;
+	grid-column-end: -2;
+	background-size: cover;
+}
