+++ frontend/src/components/Sections/HomeSection2.module.css
@@ -1,0 +1,39 @@
+.section2 {
+    height: 90vh;
+    width: 100vw;
+    background-color: aliceblue;
+}
+
+.grid_display {
+	display: grid;
+	grid-template-columns: repeat(20, 1fr);
+    grid-template-rows: repeat(20, 1fr);
+}
+
+.home_card1 {
+	grid-column-start: 3;
+	grid-column-end: 6;
+	grid-row-start: 3;
+	grid-row-end: 10;
+}
+
+.home_card2 {
+	grid-column-start: 7;
+	grid-column-end: 10;
+	grid-row-start: 4;
+	grid-row-end: 11;
+}
+
+.home_card3 {
+	grid-column-start: 3;
+	grid-column-end: 6;
+	grid-row-start: 12;
+	grid-row-end: 19;
+}
+
+.home_card4 {
+	grid-column-start: 7;
+	grid-column-end: 10;
+	grid-row-start: 13;
+	grid-row-end: 20;
+}
