--- frontend/src/components/UI/Section1.js/SectionHeading.js
@@ -1,11 +1,0 @@
-import styles from "./SectionHeading.module.css";
-
-function HeadingSection1() {
-	return (
-		<h1 className={styles.heading_section1}>
-			Lowda Ipsum ka Heading isliye ignore kro ise
-		</h1>
-	);
-}
-
-export default HeadingSection1;
