+++ frontend/src/components/UI/Section1/SectionHeading.module.css
@@ -1,0 +1,8 @@
+.heading_section1 {
+    grid-column-start: 2;
+    grid-column-end: 7;
+    grid-row: 3;
+    grid-row-end: 6;
+    color: #b8c1c6;
+    font-size: 3rem;
+}
