+++ frontend/src/components/UI/Section1/TagLine.js
@@ -1,0 +1,7 @@
+import styles from "./TagLine.module.css";
+
+function TaglLineSection1() {
+	return <h3 className={styles.taglinesection1}>Lawde ka ipsum tag line </h3>;
+}
+
+export default TaglLineSection1;
