+++ frontend/src/components/UI/Section1/TagLine.module.css
@@ -1,0 +1,8 @@
+.taglinesection1 {
+    grid-row-start: 6;
+    grid-row-end: 8;
+    grid-column-start: 2;
+    grid-column-end: 7;
+    color: #f1f6f9;
+    font-size: 1.7rem;
+}
