+++ frontend/src/components/UI/Section2/HomeCard.js
@@ -1,0 +1,17 @@
+import styles from "./HomeCard.module.css";
+
+function HomeCard(props) {
+	return (
+		<div className={`${styles.home_card} + ${props.designClass}`}>
+			<h2>{props.heading}</h2>
+			<p>{props.details}</p>
+		</div>
+	);
+}
+
+HomeCard.defaultProps = {
+    heading: "Lorem Heading",
+    details: "Loda ipsum ki details as always"
+}
+
+export default HomeCard;
